# src/DTO/CreateUserRequestDTO.ts
/**
 * @example {
 *  "name": "Cloud Strife",
 *  "email": "cloud.strife@gmail.com",
 *  "pictureUrl": "http://www.picture.com.br/cloud"
 * }
 */
export interface CreateUserRequestDTO {
  name: string;
  email: string;
  pictureUrl: string
}