# src/DTO/UpdateUserDTO.ts
/**
 * @example {
 *  "name": "Cloud Strife",
 *  "email": "cloud.strife@gmail.com",
 *  "pictureUrl": "http://www.picture.com.br/cloud"
 * }
 */
export interface UpdateUserDTO {
  name?: string;
  pictureUrl?: string
}